Namespace classes used by CSVProcessor
