+++ lib/bulk_processor/csv_processor/no_op_handler.rb
@@ -1,0 +1,15 @@
+class BulkProcessor
+  class CSVProcessor
+    # A null object implementation of the Handler role
+    class NoOpHandler
+      def initialize(payload:, successes:, errors:)
+      end
+
+      def complete!
+      end
+
+      def fail!(fatal_error)
+      end
+    end
+  end
+end
