+++ lib/bulk_processor/csv_processor/no_op_post_processor.rb
@@ -1,0 +1,16 @@
+class BulkProcessor
+  class CSVProcessor
+    # A null object implementation of the PostProcessor role
+    class NoOpPostProcessor
+      def initialize(row_processors)
+      end
+
+      def start
+      end
+
+      def errors
+        []
+      end
+    end
+  end
+end
