--- lib/bulk_processor/no_op_post_processor.rb
@@ -1,14 +1,0 @@
-class BulkProcessor
-  # A null object implementation of the PostProcessor role
-  class NoOpPostProcessor
-    def initialize(row_processors)
-    end
-
-    def start
-    end
-
-    def errors
-      []
-    end
-  end
-end
