+++ spec/bulk_processor/csv_processor/no_op_handler_spec.rb
@@ -1,0 +1,3 @@
+describe BulkProcessor::CSVProcessor::NoOpHandler do
+  it_behaves_like 'a role', 'Handler'
+end
