+++ spec/bulk_processor/csv_processor/no_op_post_processor_spec.rb
@@ -1,0 +1,3 @@
+describe BulkProcessor::CSVProcessor::NoOpPostProcessor do
+  it_behaves_like 'a role', 'PostProcessor'
+end
