--- spec/bulk_processor/no_op_handler_spec.rb
@@ -1,3 +1,0 @@
-describe BulkProcessor::NoOpHandler do
-  it_behaves_like 'a role', 'Handler'
-end
