--- spec/bulk_processor/no_op_post_processor_spec.rb
@@ -1,3 +1,0 @@
-describe BulkProcessor::NoOpPostProcessor do
-  it_behaves_like 'a role', 'PostProcessor'
-end
